# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "contact": {
      "name": "Ploiu",
      "url": "https://github.com/ploiu"
    },
    "title": "Ploiu File Server",
    "description": "a self-hostable file server written in rust",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/password": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createPassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "sets the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "201": {
            "description": "password created"
          },
          "400": {
            "description": "password has already been set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "description": "error occurred setting password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updatePassword"
              }
            }
          }
        },
        "tags": [
          "api"
        ],
        "description": "updates the password to access the api. All non /api endpoints will not work unless a password is set",
        "responses": {
          "204": {
            "description": "password updated"
          },
          "401": {
            "description": "Failed to set password. Check server logs for details"
          }
        }
      }
    },
    "/files": {
      "post": {
        "tags": [
          "files"
        ],
        "description": "upload file",
        "parameters": [
          {
            "in": "query",
            "name": "force",
            "description": "force overwrite any file with the same name",
            "example": "?force",
            "schema": {}
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/createFileRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/fileMetadataResponse"
          },
          "400": {
            "description": "missing info in the request, or a file with that name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "parent folder not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "description": "failed to save file to disk or db",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "files"
        ],
        "description": "update file metadata and location",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/fileApi"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/fileMetadataResponse"
          },
          "400": {
            "description": "a file with the same name already exists in the specified folder",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "the target file, or the destination folder, is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
...
    "/files/metadata/{id}": {
      "parameters": [
        {
          "name": "id",
          "required": true,
          "description": "the id of the file to retrieve",
          "schema": {
            "type": "number"
          },
          "in": "path"
        }
      ],
      "get": {
        "description": "retrieves the metadata of the file with the passed id",
        "tags": [
          "files"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/fileMetadataResponse"
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "description": "failed to retrieve file info. Check server logs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          }
        }
      }
    },
    "/files/{id}": {
      "parameters": [
        {
          "name": "id",
          "required": true,
          "schema": {
            "type": "number"
          },
          "in": "path"
        }
      ],
      "get": {
        "tags": [
          "files"
        ],
        "description": "downloads the file with the passed id",
        "responses": {
          "200": {
            "description": "file found and contents returned",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "$ref": "#/components/schemas/file"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "files"
        ],
        "description": "delete the file from the disk with the passed id",
        "responses": {
          "204": {
            "description": "file successfully deleted"
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
    "/files/preview/{id}": {
      "parameters": [
        {
          "name": "id",
          "required": true,
          "schema": {
            "type": "number"
          },
          "in": "path"
        }
      ],
      "get": {
        "tags": [
          "files"
        ],
        "description": "downloads a small png of the file with the passed id",
        "responses": {
          "200": {
            "description": "file found and preview returned",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/unauthorizedResponse"
          },
          "404": {
            "description": "file not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/basicMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/genericErrorResponse"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "localhost",
      "url": "https://localhost:8000"
    }
  ],
  "components": {
    "schemas": {
...
      "createFileRequest": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          },
          "extension": {
            "type": "string",
            "example": ".png"
          },
          "folderId": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          }
        }
      },
      "fileApi": {
        "properties": {
          "id": {
            "type": "number",
            "nullable": false,
            "minimum": 0
          },
          "folderId": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": false
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/tagApi"
            }
          },
          "size": {
            "type": "number",
            "nullable": true,
            "description": "never null in responses, can be null in requests (not modifiable)"
          },
          "dateCreated": {
            "type": "string",
            "nullable": true,
            "description": "never null in responses, can be null in requests (not modifiable)",
            "format": "yyyy-MM-dd'T'HH:mm:ss"
          },
          "fileType": {
            "type": "string",
            "nullable": true,
            "description": "never null in responses, can be null in requests (not modifiable)"
          }
        }
      },
...
      "basicMessage": {
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "file": {
        "type": "string",
        "format": "binary"
      },
...
      "tagApi": {
        "properties": {
          "id": {
            "nullable": true,
            "type": "number",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "createPassword": {
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          }
        }
      },
      "updatePassword": {
        "properties": {
          "oldAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          },
          "newAuth": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "nullable": false
              },
              "password": {
                "type": "string",
                "nullable": false
              }
            }
          }
        }
      }
    },
    "responses": {
      "unauthorizedResponse": {
        "description": "missing or invalid password",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/basicMessage"
            }
          }
        }
      },
      "genericErrorResponse": {
        "description": "something failed on the db or file system side of the operation. Check server logs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/basicMessage"
            }
          }
        }
      },
      "fileMetadataResponse": {
        "description": "represents a file's metadata in the server",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/fileApi"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}